requirements.txt
fastapi
uvicorn
openai
pytesseract
python-multipart
pillow
PyMuPDF
passlib[bcrypt]
python-jose[cryptography]
